+++ src/service/restRequests/campaign.tsx
@@ -1,0 +1,18 @@
+import { KeycloakInstance } from 'keycloak-js'
+import { useKeycloak } from '@react-keycloak/ssr'
+import { AxiosResponse } from 'axios'
+
+import { apiClient } from 'service/apiClient'
+import { authConfig } from 'service/restRequests'
+import { endpoints } from 'service/apiEndpoints'
+import { CampaignResponse, CampaignInput } from 'gql/campaigns'
+
+export const useCreateCampaign = () => {
+  const { keycloak } = useKeycloak<KeycloakInstance>()
+  return async (data: CampaignInput) =>
+    await apiClient.post<CampaignInput, AxiosResponse<CampaignResponse>>(
+      endpoints.campaign.createCampaign.url,
+      data,
+      authConfig(keycloak?.token),
+    )
+}
