+++ src/service/restRequests/donation.tsx
@@ -1,0 +1,12 @@
+import { AxiosResponse } from 'axios'
+
+import { apiClient } from 'service/apiClient'
+import { endpoints } from 'service/apiEndpoints'
+import { CheckoutSessionInput, CheckoutSessionResponse } from 'gql/donations'
+
+export const createCheckoutSession = async (data: CheckoutSessionInput) => {
+  return await apiClient.post<CheckoutSessionInput, AxiosResponse<CheckoutSessionResponse>>(
+    endpoints.donation.createCheckoutSession.url,
+    data,
+  )
+}
